BAH-4721 | Refactor. Rename Medication Input Control Component As MedicationRequest
